Split tests.py
